refactor(chat): extract websocket control handler
